Rename bid->ask_r, settled->go, dropped->drop
While internally it is clearer to refer to a bid and an ask, externally
it is clearer to refer to an ask and an ask_r. This makes it clear that
one operation is the reverse of the other and does not complicate the
API with the false concept of buying and selling.

To reflect this change go replaces settled and dropped becomes drop to
match the tense.

Changes the order of queue specs so that ask queue is before the
bid/ask_r queue in sbroker:start_link/3,4.