Monitor on async_ask/1 and add async_ask/2
Monitoring allows sharing a reference for monitor and reply. A reference
can be supplied as second argument (in place of the monitor).

It is unlikely that callers of both async_ask/1 and async_ask_r/1 can
both use the supervision tree to exit on broker exit. Therefore provide
a sane default (monitor), with the option not to monitor.

cancel/2 does not demonitor.