Allow 0 target in codel and 0 timeout in timeout
Allow target of 0 as this is the desired sojourn time in a LIFO queue.
Also it means both codel-like implementations, squeue_codel and
svalve_codel_r, use the same argument types.

Nolonger allow an interval of 0 as the timeout manager is more
efficient, and likely provides the desired behaviour in this case.

Allow 0 timeouts in timeout queues.