Introduce squeue:in/4 and svalve:in/4
Allow items to be enqueued before the current time.

Introduce tail_time, the time of the last insert into the tail. Use this
time as the minimum time (instead of the current time) an item is
inserted with. However active queue management will still use the current
time of the queue.

This allows a process' message queue to become an extension of an
squeue or svalve, as the time can be recorded before the sending the
message.