Improve accuracy of sregulators valve
Use the signal send time, rather than the current time, for sregulator
signals. Previously using the current time, consecutive events would
have a different time gap in the valve than the events occured. This
meant that the codel_r valve would lose accuracy under load.

When dequeuing with a valve feedback loop use the time of the signal to
give a RelativeTime - with no assumptions. It is the time the signal was
sent minus the time the ask request was sent, as with sbroker's
RelativeTime.

When the regulator dequeues an item due to going below the minimum
threshold the internal queue sojourn time is still used for the
RelativeTime.