Control timeout with queue instead of fixed timeout
Replace the fixed timeout return by an `sbroker` `init/1` callback with
a timeout set by the active `sbroker_queue`. This provides more accurate
queue management as the active queue can cause an item to be dropped as
soon as desired but returning a suitable timeout. It also allows a
broker to go completely idle when both queues are empty by using an
infinity timeout.

* `sbroker` `init/1` callback now returns {ok, {QueueSpec, QueueSpec}}
* Most `sbroker_queue` callbacks must return a `TimeoutNext` value
* `sbroker_time:convert_time_unit/3` introduced to match `erlang`'s