Improve relative time in meters/regulator
Previously sbroker and sregulator would use the dequeue time of the
last dequeue or open time, which would be inaccurate representation of
congestion. Instead ask the active queue/valve for its send/open time.

Also run the valve at the send time instead of actual time so that
relative time behaves as if the valve was a separate process and
relative time accurate represents congestion. This also means that a
batch of update/casts will be handled at the time sent rather than
spread out over the handled time.